cupid: overlay: Increase auto brightness debounce time
Our light sensor is getting confused when the screen brightness changes
in a rather dark environment and sometimes jumps between 0 and a higher
brightness leading to a raise in brightness, but then the sensor
over-corrects the read brightness and reports 0 again. Try to fix this
by giving the sensor some more time to debounce.

Change-Id: Ice8ad0b4e34af16bc905f84c27bd491378617cae