cupid: Remove 32-bit libQnnHtpAltPrepV69Skel from vendor/lib64/camera
The file is a elf32-little file but it is copied to lib64.
Extract-utils now declares the target as android_arm due to that,
causing issues in combination with compile_mutilib being set to 64.
libQnnHtpAltPrepV69Skel is also in vendor/lib/rfsa/adsp and it's
module name is the same (which the build system seems to accept?)
which lead to the module not being present in the end since a few
weeks anyway.

Change-Id: I9a470479697a7d66af0066197ce0bbfc383f973e